Update start of 10-open.md
Shortened some of the text at the start. I think it's better to focus on the best practice workflow for a paper. Given that it's very likely none of the students will have published anything before, better not to even mention the first part. Same goes for the exercise on 'reproducing your partner's publication from the web'.